requirements.txt
# Основа
aiogram>=3.0  # Для бота (асинхронный, быстрый)
pydantic>=2.0 # Валидация данных (как в ТЗ)
python-dotenv # Чтение .env

# Для мозгов (LLM)
langchain
langchain-openai
langchain-community
openai

# Форматирование кода
ruff
black


sqlalchemy>=2.0.0
asyncpg
pgvector

selectolax==0.3.21
curl-cffi==0.7.1

markdownify
